- Don't call KSToReflectance but do the operation in line. - Fix the ChannelConverter Test
svn path=/trunk/koffice/; revision=755306

+++ krita/plugins/painterlyframework/tests/channel_converter_test.cpp
@@ -33,21 +33,21 @@ void ChannelConverterTest::testKSReflectance()
     R = 0.4999999999999;
     c.reflectanceToKS(R, K, S);
     qDebug() << "Reflectance " << R << "; K = " << K << ", S = " << S;
-    QCOMPARE((double)(S/K), (R*2.0/pow(1-R,2)));
+    QCOMPARE((S/K), (float)(R*2.0/pow(1-R,2)));
 
     R = 0.5000000000001;
     c.reflectanceToKS(R, K, S);
     qDebug() << "Reflectance " << R << "; K = " << K << ", S = " << S;
-    QCOMPARE((double)(S/K), (R*2.0/pow(1-R,2)));
+    QCOMPARE((S/K), (float)(R*2.0/pow(1-R,2)));
 
     for (int i = 0; i <= 100; i++) {
         R = (double)i/100.0;
         c.reflectanceToKS(R, K, S);
         qDebug() << "Reflectance " << R << "; K = " << K << ", S = " << S;
         if ( i <= 50 )
-            QCOMPARE((double)(S/K), (2.0*R*pow(1-R,-2)));
+            QCOMPARE((S/K), (float)(2.0*R*pow(1-R,-2)));
         else
-            QCOMPARE((double)(K/S), (0.5*pow(1-R, 2)/R));
+            QCOMPARE((K/S), (float)(0.5*pow(1-R, 2)/R));
     }
 
     // KSToReflectance
@@ -65,7 +65,7 @@ void ChannelConverterTest::testKSReflectance()
         c.KSToReflectance(K, S, R);
         qDebug() << "K " << K << ", S " << S << "; R = " << R;
         QVERIFY(R > 0.0 && R < 1.0);
-        QCOMPARE((double)(S/K), (R*2.0/pow(1-R,2)));
+        QCOMPARE((S/K), (float)(R*2.0/pow(1-R,2)));
     }
 }
